+++ src/app/[chain]/chart/marketcap/page.tsx
@@ -7,7 +7,7 @@ import { IMarkerCap } from '../type';
 import { exportToCSV } from '@_utils/urlUtils';
 import { fetchDailyMarketCap } from '@_api/fetchChart';
 import PageLoadingSkeleton from '@_components/PageLoadingSkeleton';
-import { useFetchChartData } from '@_hooks/useFetchChartData';
+import { useChartDownloadData, useFetchChartData } from '@_hooks/useFetchChartData';
 const title = 'aelf Market Cap Chart';
 const getOption = (list: any[]): Highcharts.Options => {
   const allData: any[] = [];
@@ -60,19 +60,8 @@ export default function Page() {
     return getOption(data?.list || []);
   }, [data]);
 
-  useEffect(() => {
-    if (data) {
-      const chart = chartRef.current?.chart;
-      if (chart) {
-        const minDate = data.list[0]?.date;
-        const maxDate = data.list[data.list.length - 1]?.date;
-        chart.xAxis[0].setExtremes(minDate, maxDate);
-      }
-    }
-  }, [chartRef, data]);
-  const download = () => {
-    exportToCSV(data?.list || [], title);
-  };
+  const { download } = useChartDownloadData(data, chartRef, title);
+
   const highlightData = [];
   return loading ? (
     <PageLoadingSkeleton />
