+++ src/app/[chain]/chart/price/page.tsx
@@ -7,7 +7,7 @@ import { ChartColors, IDailyPriceDData } from '../type';
 import { exportToCSV } from '@_utils/urlUtils';
 import { fetchDailyElfPrice } from '@_api/fetchChart';
 import PageLoadingSkeleton from '@_components/PageLoadingSkeleton';
-import { useFetchChartData } from '@_hooks/useFetchChartData';
+import { useChartDownloadData, useFetchChartData } from '@_hooks/useFetchChartData';
 
 const title = 'ELF Daily Price (USD) Chart';
 
@@ -47,18 +47,8 @@ export default function Page() {
   });
 
   const options = useMemo(() => getOption(data?.list || []), [data]);
-  useEffect(() => {
-    if (data) {
-      const chart = chartRef.current?.chart;
-      if (chart) {
-        const minDate = data.list[0]?.date;
-        const maxDate = data.list.at(-1)?.date;
-        chart.xAxis[0].setExtremes(minDate, maxDate);
-      }
-    }
-  }, [chartRef, data]);
 
-  const download = () => exportToCSV(data?.list || [], title);
+  const { download } = useChartDownloadData(data, chartRef, title);
 
   return loading ? (
     <PageLoadingSkeleton />
