+++ src/app/[chain]/chart/transactions/page.tsx
@@ -9,7 +9,7 @@ import { exportToCSV } from '@_utils/urlUtils';
 import { fetchDailyTransactions } from '@_api/fetchChart';
 import PageLoadingSkeleton from '@_components/PageLoadingSkeleton';
 import { useMultiChain } from '@_hooks/useSelectChain';
-import { useFetchChartData } from '@_hooks/useFetchChartData';
+import { useChartDownloadData, useFetchChartData } from '@_hooks/useFetchChartData';
 
 const title = 'aelf Daily Transactions Chart';
 const getOption = (list: any[], multi, chain): Highcharts.Options => {
@@ -90,29 +90,16 @@ const getOption = (list: any[], multi, chain): Highcharts.Options => {
   return options;
 };
 export default function Page() {
-  const { data, loading, chartRef, chain } = useFetchChartData<IDailyTransactionsData>({
+  const { data, loading, chartRef, chain, multi } = useFetchChartData<IDailyTransactionsData>({
     fetchFunc: fetchDailyTransactions,
     processData: (res) => res,
   });
-  const multi = useMultiChain();
+
   const options = useMemo(() => {
     return getOption(data?.list || [], multi, chain);
   }, [chain, data?.list, multi]);
 
-  useEffect(() => {
-    if (data) {
-      const chart = chartRef.current?.chart;
-      if (chart) {
-        const minDate = data.list[0]?.date;
-        const maxDate = data.list[data.list.length - 1]?.date;
-        chart.xAxis[0].setExtremes(minDate, maxDate);
-      }
-    }
-  }, [chartRef, data]);
-
-  const download = () => {
-    exportToCSV(data?.list || [], title);
-  };
+  const { download } = useChartDownloadData(data, chartRef, title);
 
   const highlightData = useMemo<IHIGHLIGHTDataItem[]>(() => {
     const key = multi ? 'mergeTransactionCount' : 'transactionCount';
