+++ src/app/[chain]/chart/tvl/page.tsx
@@ -7,7 +7,7 @@ import { IHIGHLIGHTDataItem, ITVLData } from '../type';
 import { exportToCSV } from '@_utils/urlUtils';
 import { fetchDailyTvl } from '@_api/fetchChart';
 import PageLoadingSkeleton from '@_components/PageLoadingSkeleton';
-import { useFetchChartData } from '@_hooks/useFetchChartData';
+import { useChartDownloadData, useFetchChartData } from '@_hooks/useFetchChartData';
 const title = 'TVL Chart';
 const getOption = (list: any[]): Highcharts.Options => {
   const allData: any[] = [];
@@ -80,19 +80,8 @@ export default function Page() {
     return getOption(data?.list || []);
   }, [data]);
 
-  useEffect(() => {
-    if (data) {
-      const chart = chartRef.current?.chart;
-      if (chart) {
-        const minDate = data.list[0]?.date;
-        const maxDate = data.list[data.list.length - 1]?.date;
-        chart.xAxis[0].setExtremes(minDate, maxDate);
-      }
-    }
-  }, [chartRef, data]);
-  const download = () => {
-    exportToCSV(data?.list || [], title);
-  };
+  const { download } = useChartDownloadData(data, chartRef, title);
+
   const highlightData = useMemo<IHIGHLIGHTDataItem[]>(() => {
     return data
       ? [
